fix 32/64 compat avoid the use of size_t on x86 architecture
Former-commit-id: fa7c98d7eab74105683481a8e8433ae59afaea4d [formerly 32c3509e87b056c17af8a893edabd80935efd11f]
Former-commit-id: 6b68bfb19c2936d0e56c2bdd77200affdb6cf455

+++ src/Acquisition.hpp
@@ -202,7 +202,7 @@ class SRC_API Acquisition
     long long m_start = 0; // microseconds
     long long m_end = 0; // microseconds
     data::Measures m_measures;
-    size_t m_size = 0;
+    uint64_t m_size = 0;
 
     friend class io::Input;
     friend class io::Output;
